+++ src/chains/Bnb/Common.sol
@@ -15,13 +15,11 @@ import {
     pancakeInfinityClManager,
     pancakeInfinityBinManager
 } from "../../core/pancakeInfinityForks/PancakeInfinity.sol";
-import {EulerSwap, IEVC, IEulerSwap} from "../../core/EulerSwap.sol";
 import {Bebop} from "../../core/Bebop.sol";
 
 import {FreeMemory} from "../../utils/FreeMemory.sol";
 
 import {ISettlerActions} from "../../ISettlerActions.sol";
-import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 import {revertUnknownForkId} from "../../core/SettlerErrors.sol";
 
 import {
@@ -60,7 +58,6 @@ abstract contract BnbMixin is
     DodoV2,
     UniswapV4,
     PancakeInfinity,
-    EulerSwap,
     Bebop
 {
     constructor() {
@@ -89,11 +86,6 @@ abstract contract BnbMixin is
             ) = abi.decode(data, (address, IERC20, uint256, bool, uint256, uint256, bytes, uint256));
 
             sellToUniswapV4(recipient, sellToken, ppm, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
-        } else if (action == uint32(ISettlerActions.EULERSWAP.selector)) {
-            (address recipient, IERC20 sellToken, uint256 ppm, IEulerSwap pool, bool zeroForOne, uint256 amountOutMin) =
-                abi.decode(data, (address, IERC20, uint256, IEulerSwap, bool, uint256));
-
-            sellToEulerSwap(recipient, sellToken, ppm, pool, zeroForOne, amountOutMin);
         } else if (action == uint32(ISettlerActions.MAVERICKV2.selector)) {
             (
                 address recipient,
@@ -192,10 +184,6 @@ abstract contract BnbMixin is
 
     function _PANCAKE_INFINITY_BIN_MANAGER() internal pure override returns (address) {
         return pancakeInfinityBinManager;
-    }
-
-    function _EVC() internal pure override returns (IEVC) {
-        return IEVC(0xb2E5a73CeE08593d1a076a2AE7A6e02925a640ea);
     }
 
     // I hate Solidity inheritance